Extract parser routines and security grammar
